break circdep in chromium
