+++ src/core/puppeteer.ts
@@ -7,9 +7,9 @@
 import { readRegistryKey } from "./windows.ts";
 import { which } from "./path.ts";
 import { error, info } from "log/mod.ts";
-import { fetcher } from "../tools/impl/chromium.ts";
 import { existsSync } from "fs/mod.ts";
 import { UnreachableError } from "./lib/error.ts";
+import { quartoDataDir } from "./appdirs.ts";
 
 // deno-lint-ignore no-explicit-any
 let puppeteerImport: any = undefined;
@@ -279,4 +279,16 @@ async function fetchBrowser() {
     product: "chrome",
     executablePath,
   });
+}
+
+export async function fetcher() {
+  const options = {
+    path: chromiumInstallDir(),
+  };
+  const fetcher = (await getPuppeteer()).createBrowserFetcher(options);
+  return fetcher;
+}
+
+export function chromiumInstallDir(): string | undefined {
+  return quartoDataDir("chromium");
 }
